server/requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
python-dotenv>=1.0.1
openai>=1.99.0
yt-dlp>=2025.1.0

